pyproject.toml
[tool.poetry]
name = "demoblaze-tests"
version = "0.1.0"
description = ""
authors = ["lrayne <legaterayne@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selene = {version = "^2.0.0rc9", allow-prereleases = true}
pytest = "^8.2.2"
black = "^24.4.2"
python-dotenv = "^1.0.1"
allure-pytest = "^2.13.5"
mimesis = "^17.0.0"

[tool.black]
line-length = 79
skip-string-normalization = 1

[tool.pytest.ini_options]
addopts = [
    '--clean-alluredir',
    '--alluredir=allure-results'
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
